+++ R/longer_wider.R
@@ -1,22 +1,22 @@
 #' @export
-pivot_longer <- function(.data, ...) {
-    log_longer_wider(.data, .fun = tidyr::pivot_longer, .funname = "pivot_longer", ...)
+pivot_longer <- function(data, ...) {
+    log_longer_wider(data, .fun = tidyr::pivot_longer, .funname = "pivot_longer", ...)
 }
 
 #' @export
-pivot_wider <- function(.data, ...) {
-    log_longer_wider(.data, .fun = tidyr::pivot_wider, .funname = "pivot_wider", ...)
+pivot_wider <- function(data, ...) {
+    log_longer_wider(data, .fun = tidyr::pivot_wider, .funname = "pivot_wider", ...)
 }
 
 #' @export
-gather <- function(.data, ...) {
-    log_longer_wider(.data, .fun = tidyr::gather, .funname = "gather", ...)
+gather <- function(data, ...) {
+    log_longer_wider(data, .fun = tidyr::gather, .funname = "gather", ...)
 }
 
 
 #' @export
-spread <- function(.data, ...) {
-    log_longer_wider(.data, .fun = tidyr::spread, .funname = "spread", ...)
+spread <- function(data, ...) {
+    log_longer_wider(data, .fun = tidyr::spread, .funname = "spread", ...)
 }
 
 log_longer_wider <- function(.data, .fun, .funname, ...) {